Successfully move construct_runtime to a separate file
